+++ Source/WebCore/loader/FrameLoader.cpp
@@ -668,7 +668,7 @@ void FrameLoader::finishedParsing()
     // Null-checking the FrameView indicates whether or not we're in the destructor.
     RefPtr<Frame> protector = m_frame->view() ? m_frame : 0;
 
-    m_client->dispatchDidFinishDocumentLoad();
+    m_client->dispatchDidFinishDocumentLoad(this);
 
     checkCompleted();